Split buffers into their own tests
