+++ crates/test-rust-wasm/src/bin/buffers.rs
@@ -1,0 +1,3 @@
+include!("../../../../tests/runtime/buffers/wasm.rs");
+
+fn main() {}
