+++ tests/runtime/wasmtime-run/imports.rs
@@ -80,111 +80,6 @@ impl Host for MyHost {
         (b"typedef3".to_vec(), vec!["typedef4".to_string()])
     }
 
-    fn buffer_u8(&mut self, in_: &[u8], out: &mut [u8]) -> u32 {
-        assert_eq!(in_, [0]);
-        assert_eq!(out.len(), 10);
-        out[0] = 1;
-        out[1] = 2;
-        out[2] = 3;
-        3
-    }
-
-    fn buffer_u32(&mut self, in_: &[Le<u32>], out: &mut [Le<u32>]) -> u32 {
-        assert_eq!(in_.len(), 1);
-        assert_eq!(in_[0].get(), 0);
-        assert_eq!(out.len(), 10);
-        out[0].set(1);
-        out[1].set(2);
-        out[2].set(3);
-        3
-    }
-
-    fn buffer_bool(&mut self, in_: PullBuffer<'_, bool>, mut out: PushBuffer<'_, bool>) -> u32 {
-        assert!(in_.len() <= out.capacity());
-        let len = in_.len();
-        for item in in_.iter() {
-            let item = item.unwrap();
-            out.write(Some(!item)).unwrap();
-        }
-        len as u32
-    }
-
-    // fn buffer_string(
-    //     &mut self,
-    //     in_: PullBuffer<'_, GuestPtr<'_, str>>,
-    //     mut out: PushBuffer<'_, String>,
-    // ) -> u32 {
-    //     assert!(in_.len() < out.capacity());
-    //     let len = in_.len();
-    //     for item in in_.iter().unwrap() {
-    //         let item = item.unwrap();
-    //         let s = item.borrow().unwrap();
-    //         out.write(Some(s.to_uppercase())).unwrap();
-    //     }
-    //     len as u32
-    // }
-
-    // fn buffer_list_bool(
-    //     &mut self,
-    //     in_: PullBuffer<'_, Vec<bool>>,
-    //     mut out: PushBuffer<'_, Vec<bool>>,
-    // ) -> u32 {
-    //     assert!(in_.len() < out.capacity());
-    //     let len = in_.len();
-    //     for item in in_.iter().unwrap() {
-    //         let item = item.unwrap();
-    //         out.write(Some(item.into_iter().map(|b| !b).collect()))
-    //             .unwrap();
-    //     }
-    //     len as u32
-    // }
-
-    // fn buffer_buffer_bool(&mut self, in_: PullBuffer<'_, PullBuffer<'_, bool>>) {
-    //     assert_eq!(in_.len(), 1);
-    //     let buf = in_.iter().unwrap().next().unwrap().unwrap();
-    //     assert_eq!(buf.len(), 5);
-    //     assert_eq!(
-    //         buf.iter()
-    //             .unwrap()
-    //             .collect::<Result<Vec<bool>, _>>()
-    //             .unwrap(),
-    //         [true, false, true, true, false]
-    //     );
-    // }
-
-    fn buffer_mutable1(&mut self, a: Vec<PullBuffer<'_, bool>>) {
-        assert_eq!(a.len(), 1);
-        assert_eq!(a[0].len(), 5);
-        assert_eq!(
-            a[0].iter().collect::<Result<Vec<_>, _>>().unwrap(),
-            [true, false, true, true, false]
-        );
-    }
-
-    fn buffer_mutable2(&mut self, mut a: Vec<&mut [u8]>) -> u32 {
-        assert_eq!(a.len(), 1);
-        assert!(a[0].len() > 4);
-        a[0][..4].copy_from_slice(&[1, 2, 3, 4]);
-        return 4;
-    }
-
-    fn buffer_mutable3(&mut self, mut a: Vec<PushBuffer<'_, bool>>) -> u32 {
-        assert_eq!(a.len(), 1);
-        assert!(a[0].capacity() > 3);
-        a[0].write([false, true, false].iter().copied()).unwrap();
-        return 3;
-    }
-
-    fn buffer_in_record(&mut self, _: BufferInRecord<'_>) {}
-    fn buffer_typedef(
-        &mut self,
-        _: ParamInBufferU8<'_>,
-        _: ParamOutBufferU8<'_>,
-        _: ParamInBufferBool<'_>,
-        _: ParamOutBufferBool<'_>,
-    ) {
-    }
-
     fn list_of_variants(
         &mut self,
         bools: Vec<bool>,
